zest_cli/requirements.txt
llama-cpp-python>=0.2.0
requests>=2.31.0
